requirements.txt
pandas>=1.5.0
faker==8.10.0
pyspark==3.1.2
findspark==1.4.2
google-cloud-bigquery>=3.27.0
pyarrow>=18.0.0
